+++ backend/src/main/java/com/tibame/app_generator/AppGeneratorBackendApplication.java
@@ -1,9 +1,12 @@
 package com.tibame.app_generator;
 
+import com.tibame.app_generator.config.StorageProperties;
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
+import org.springframework.boot.context.properties.EnableConfigurationProperties;
 
 @SpringBootApplication
+@EnableConfigurationProperties(StorageProperties.class)
 public class AppGeneratorBackendApplication {
 
 	public static void main(String[] args) {